+++ internal/riscv64/bytes_rvv_riscv64.go
@@ -124,6 +124,9 @@ func countAnyRVV(b []byte, chars []byte) int
 func jsonMasksRVV(dst []byte, b []byte, want uint32)
 
 //go:noescape
+func jsonValidTokensRVV(b []byte, masks []uint64, stk []uint64) int
+
+//go:noescape
 func jsonStage1RVV(out []uint64, masks []byte, nw int, carr []uint64, res []int64)
 
 //go:noescape